+++ dgemm/util.h
@@ -4,4 +4,4 @@
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 
 void assert_result(int n, double m[n][n], double expected);
-void print_performance(double elapsed_time, unsigned long flps);
+void print_performance(double elapsed_time, unsigned long long flps);
